enh: Turn project into MIRTK Package providing evaluate-slice-dice command
