+++ config/Depends.cmake
@@ -1,0 +1,3 @@
+if (MIRTK_Image_WITH_VTK)
+  basis_find_package(VTK 7|6 REQUIRED COMPONENTS vtkCommonCore vtkCommonDataModel)
+endif ()
